Update task if conf changes, and restart always
We will get supervisor to update itself if the config has changed.
This will force it to restart with the updated commands.

Also, after much deliberation, I think that it should always restart the
supervisor task. If the task is delicate and can't be restarted, perhaps
being a supervisor task isn't the best idea.