Always update supervisor
Using the notify handler doesn't always work. In particular, it looks
like the first time the file is created, it doesn't trigger the handler!
This is disastrous as it means that when the role tries to restart the
supervisor task, it fails :(

So we always call 'supervisor update'. It is the only reliable way I can
think of.